% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known-sources.R
\docType{data}
\name{all_sources}
\alias{all_sources}
\title{Known data files}
\format{
An object of class \code{list} of length 13.
}
\usage{
all_sources
}
\description{
Known data files
}
\keyword{datasets}
