% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-array-altrep-materialization.R
\docType{data}
\name{array_altrep_materialization}
\alias{array_altrep_materialization}
\title{Benchmark for materializing an altrep Arrow array}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
array_altrep_materialization
}
\description{
This flexes a lower level conversion to R data structures from Arrow data structures.
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id to use (it will be read in to a data.frame first)
\item \code{exclude_nulls} logical, should any columns with any \code{NULL}s or \code{NA}s in them be removed?
}
}

\keyword{datasets}
