% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-array-to-vector.R
\docType{data}
\name{array_to_vector}
\alias{array_to_vector}
\title{Benchmark for reading an Arrow table to a data.frame}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
array_to_vector
}
\description{
This flexes that conversion to R data structures from Arrow data structures.
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id to use (it will be read in to a data.frame first)
}
}

\keyword{datasets}
