% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.data.frame.BenchmarkResults}
\alias{as.data.frame.BenchmarkResults}
\alias{as.data.frame.BenchmarkResult}
\title{Convert benchmark result object to a tidy data frame}
\usage{
\method{as.data.frame}{BenchmarkResults}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{BenchmarkResult}(x, row.names = NULL, optional = FALSE, packages = "arrow", ...)
}
\arguments{
\item{x}{a benchmark result object or list of them as returned by \code{\link[=run_one]{run_one()}} or \code{\link[=run_benchmark]{run_benchmark()}}}

\item{row.names}{for generic consistency}

\item{optional}{for generic consistency}

\item{...}{additional arguments passed on to methods for individual results.
\code{packages} is the only currently supported argument.}

\item{packages}{Packages for which to extract versions}
}
\value{
A data.frame suitable for analysis in R
}
\description{
Convert benchmark result object to a tidy data frame
}
