% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.data.frame.conbench_results}
\alias{as.data.frame.conbench_results}
\title{Convert benchmark result object to a tidy data frame}
\usage{
\method{as.data.frame}{conbench_results}(results, ...)
}
\arguments{
\item{results}{list of benchmark results returned by \code{\link[=run_benchmark]{run_benchmark()}}}

\item{...}{additional arguments passed to \code{.result_to_df()}. \code{packages} is
the only currently supported argument, which governs which package versions
to keep as columns in the result.}
}
\value{
A data.frame suitable for analysis in R
}
\description{
Convert benchmark result object to a tidy data frame
}
