% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{assemble_metadata}
\alias{assemble_metadata}
\title{Assemble metadata for a benchmark run}
\usage{
assemble_metadata(name, params, cpu_count, n_iter)
}
\arguments{
\item{name}{Benchmark name, i.e. \code{bm$name}}

\item{params}{Named list of parameters for the individual run, i.e. the case}

\item{cpu_count}{Number of CPUs allocated}

\item{n_iter}{Number of iterations}
}
\description{
Assemble metadata for a benchmark run
}
\keyword{internal}
