% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{augment_result}
\alias{augment_result}
\title{Augment a benchmark result with collected defaults}
\usage{
augment_result(result)
}
\arguments{
\item{result}{An instance of \link{BenchmarkResult}}
}
\description{
Augment a benchmark result with collected defaults
}
