% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{BenchEnvironment}
\alias{BenchEnvironment}
\title{Create a test environment to run benchmarks in}
\usage{
BenchEnvironment(...)
}
\arguments{
\item{...}{named list of parameters to set in the environment}
}
\value{
An environment
}
\description{
Create a test environment to run benchmarks in
}
