% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{Benchmark}
\alias{Benchmark}
\title{Define a Benchmark}
\usage{
Benchmark(
  setup = function(ctx, ...) { },
  before_each = function(ctx) { },
  run = function(ctx) { },
  after_each = function(ctx) { },
  teardown = function(ctx) { },
  valid_params = function(params) params,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
A \code{Benchmark} object containing these functions
}
\description{
Define a Benchmark
}
\section{Evaluation}{

A \code{Benchmark} is evaluated something like:\preformatted{ctx <- new.env()
bm$setup(ctx, param1 = "value", param2 = "value")
for (i in seq_len(n_iter)) \{
  bm$before_each(ctx)
  measure(bm$run(ctx))
  bm$after_each(ctx)
\}
bm$teardown(ctx)
}
}

