% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-dataframe.R
\name{BenchmarkDataFrame}
\alias{BenchmarkDataFrame}
\title{A classed dataframe of benchmarks for running}
\usage{
BenchmarkDataFrame(benchmarks, parameters)
}
\arguments{
\item{benchmarks}{A list with elements of class \code{Benchmark}}

\item{parameters}{Optional. A list of dataframes of parameter combinations to
run as generated by \code{\link[=get_default_parameters]{get_default_parameters()}}. If null, defaults will be generated
when \code{\link[=run]{run()}} is called.}
}
\value{
A classed dataframe with \code{name} (benchmark attribute, not object name),
\code{benchmark}, and \code{params} columns
}
\description{
A classed dataframe of benchmarks for running
}
