% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{confirm_mem_alloc}
\alias{confirm_mem_alloc}
\title{Confirm that the memory allocator enabled}
\usage{
confirm_mem_alloc(mem_alloc)
}
\arguments{
\item{mem_alloc}{the memory allocator to be tested (one of: "jemalloc", "mimalloc", "system")}
}
\value{
nothing
}
\description{
Confirm that the memory allocator enabled
}
\keyword{internal}
