% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{data_file}
\alias{data_file}
\title{Find a data file}
\usage{
data_file(...)
}
\arguments{
\item{...}{file path to look for}
}
\value{
path to the file
}
\description{
This looks in both the source dir ("data") as well as the temp directory
("data/temp") and returns the temp directory file if it exists, if it doesn't
but the source dir exists it will return that, if neither exist it returns
the temp directory path.
}
\keyword{internal}
