% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalogistik.R
\name{datalogistik_get}
\alias{datalogistik_get}
\title{Call \verb{datalogistik get}}
\usage{
datalogistik_get(
  dataset,
  format = NULL,
  compression = NULL,
  scale_factor = NULL
)
}
\arguments{
\item{dataset}{the dataset to get}

\item{format}{the format of the dataset to be in (default: parquet)}

\item{compression}{the compression of the dataset to use (default: uncompressed)}

\item{scale_factor}{the scale_factor of the dataset (for datasets that use it)}
}
\value{
The returned metadata JSON parsed into a list
}
\description{
Call \verb{datalogistik get}
}
