% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-dataset-write.R
\docType{data}
\name{dataset_repartition}
\alias{dataset_repartition}
\title{Benchmark Taxi dataset (Parquet) re-partitioning}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
dataset_repartition
}
\description{
Benchmark Taxi dataset (Parquet) re-partitioning
}
\section{Parameters}{

\itemize{
\item \code{columns} Names of new partition keys
}
}

\keyword{datasets}
