% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-dataset-taxi-2013.R
\docType{data}
\name{dataset_taxi_2013}
\alias{dataset_taxi_2013}
\title{Benchmark Taxi 2013 dataset reading}
\format{
An object of class \code{Benchmark} of length 12.
}
\usage{
dataset_taxi_2013
}
\description{
Benchmark Taxi 2013 dataset reading
}
\section{Parameters}{

\itemize{
\item \code{dataset} Name of dataset to use, either \code{taxi_2013} or \code{taxi_2013_sample} (for testing)
\item \code{query} Name of a known query to run; see \code{dataset_taxi_2013$cases}
}
}

\keyword{datasets}
