% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-dataset-taxi-parquet.R
\docType{data}
\name{dataset_taxi_parquet}
\alias{dataset_taxi_parquet}
\title{Benchmark Taxi dataset (Parquet) reading}
\format{
An object of class \code{Benchmark} of length 10.
}
\usage{
dataset_taxi_parquet
}
\description{
Benchmark Taxi dataset (Parquet) reading
}
\section{Parameters}{

\itemize{
\item \code{query} Name of a known query to run; see \code{dataset_taxi_parquet$cases}
}
}

\keyword{datasets}
