% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{default_params}
\alias{default_params}
\title{Generate a dataframe of default parameters for a benchmark}
\usage{
default_params(x, ...)
}
\arguments{
\item{x}{An object for which to generate parameters}

\item{...}{Named arguments corresponding to the parameters of \code{bm}'s \code{setup}
function. May also contain \code{cpu_count}, \code{lib_path}, and \code{mem_alloc}.}
}
\value{
For \code{default_params.Benchmark}, a dataframe of parameter combinations
to try with a column for each parameter and a row for each combination.
}
\description{
Generates a dataframe of parameter combinations for a benchmark to try based
on the parameter defaults of its \code{setup} function and supplied parameters.
}
