% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-df-to-table.R
\docType{data}
\name{df_to_table}
\alias{df_to_table}
\title{Benchmark for reading a data.frame into an Arrow table}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
df_to_table
}
\description{
This flexes that conversion from R data structures to Arrow data structures.
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id to use (it will be read in to a data.frame first)
}
}

\keyword{datasets}
