% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-format.R
\name{ensure_format}
\alias{ensure_format}
\title{Ensure that a source has a specific format}
\usage{
ensure_format(
  name,
  format = known_formats,
  compression = known_compressions,
  num_groups = NULL
)
}
\arguments{
\item{name}{name of the known source}

\item{format}{format to be ensured}

\item{compression}{compression to be ensured}

\item{num_groups}{the number of groups to split a file into (default: NULL, which uses the default)}
}
\value{
the file that was ensured to exist
}
\description{
Ensure that a source has a specific format
}
