% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{ensure_source}
\alias{ensure_source}
\title{Ensure that a source has a specific format}
\usage{
ensure_source(
  name,
  format = known_formats,
  compression = known_compressions,
  chunk_size = NULL,
  scale_factor = NULL
)
}
\arguments{
\item{name}{name of the known source}

\item{format}{format to be ensured}

\item{compression}{compression to be ensured}

\item{chunk_size}{the number of rows to write in each chunk}

\item{scale_factor}{the scale factor to use (for datasources where that is relevant)}
}
\value{
the file that was ensured to exist
}
\description{
Ensure that a source has a specific format
}
