% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-tpch-source.R
\name{ensure_tpch}
\alias{ensure_tpch}
\title{Generate tpch data}
\usage{
ensure_tpch(scale_factor = 1)
}
\arguments{
\item{scale_factor}{a relative measure of the size of data in gigabytes.}
}
\description{
Generate tpch data at a given scale factor. By default,
data is output relative to the current working directory. However,
you can set the environment variable \code{ARROWBENCH_DATA_DIR} to
point to another directory. Setting this environment variable has
the advantage of being a central location for general usage. Running
this function will install a custom version of duckdb in an \code{r_libs}
directory, relative to the directory specified by the environment
variable \code{ARROWBENCH_LOCAL_DIR}. When running this function for the first time you will
see significant output from that installation process. This is normal.
}
