% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{file_with_ext}
\alias{file_with_ext}
\title{Get a file with an extension}
\usage{
file_with_ext(file, new_ext)
}
\arguments{
\item{file}{the file}

\item{new_ext}{the new extension}
}
\value{
the file with the new extension
}
\description{
Get a file with an extension
}
\keyword{internal}
