% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-tpch-source.R
\name{generate_tpch}
\alias{generate_tpch}
\title{Generate tpch data}
\usage{
generate_tpch(scale_factor = 1)
}
\arguments{
\item{scale_factor}{a relative measure of the size of data in gigabytes.}
}
\description{
Generate tpch data at a given scale factor. By default,
\code{data} is assumed to be relative to the current working directory, but
you can set the environment variable \code{ARROWBENCH_DATA_DIR} to point to another
(permanent) base directory.
}
