% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-csv.R
\name{get_csv_reader}
\alias{get_csv_reader}
\title{Get a CSV reader}
\usage{
get_csv_reader(reader, delim)
}
\arguments{
\item{reader}{the reader to use}

\item{delim}{the delimiter to use}
}
\value{
the csv reader
}
\description{
Get a CSV reader
}
