% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-write-csv.R
\name{get_csv_writer}
\alias{get_csv_writer}
\title{Get a CSV writer}
\usage{
get_csv_writer(writer)
}
\arguments{
\item{writer}{the writer to use}
}
\value{
the csv writer
}
\description{
Get a CSV writer
}
