% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{get_dataset_attr}
\alias{get_dataset_attr}
\title{Get dataset attributes}
\usage{
get_dataset_attr(name, attr)
}
\arguments{
\item{attr}{the attribute to get}

\item{dataset}{the file to get attributes for}
}
\description{
Get dataset attributes
}
\keyword{internal}
