% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{get_input_func}
\alias{get_input_func}
\title{Get an input function for a table}
\usage{
get_input_func(
  engine,
  scale_factor,
  query_id,
  format,
  compression = "uncompressed",
  con = NULL,
  memory_map = FALSE,
  num_groups = parallel::detectCores()
)
}
\arguments{
\item{engine}{which engine to use}

\item{scale_factor}{what scale factor to reference}

\item{query_id}{which query is being used}

\item{format}{which format}

\item{compression}{which compression to use (default: "uncompressed")}

\item{con}{a connection}

\item{memory_map}{should the file be memory mapped? (only relevant for the "native" format with Arrow)}
}
\description{
This returns a function which will return a table reference with the specified
parameters
}
