% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-json.R
\name{get_json_reader}
\alias{get_json_reader}
\title{Get a JSON reader}
\usage{
get_json_reader(reader)
}
\arguments{
\item{reader}{string of the reader package to use}
}
\value{
the JSON function
}
\description{
Get a JSON reader
}
