% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-dataframe.R
\name{get_package_benchmarks}
\alias{get_package_benchmarks}
\title{Get a list of benchmarks in a package}
\usage{
get_package_benchmarks(package = "arrowbench")
}
\arguments{
\item{package}{String of package name in which to find benchmarks}
}
\value{
An instance of \link{BenchmarkDataFrame} with all the benchmarks contained
by a package
}
\description{
Get a list of benchmarks in a package
}
