% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{get_query_func}
\alias{get_query_func}
\title{Get a query function that will run a specific TPC-H query}
\usage{
get_query_func(query_id, engine = NULL)
}
\arguments{
\item{query_id}{which query to get?}

\item{engine}{which engine to use (all options return a dplyr-based query,
with the except of \code{"duckdb_sql"} which returns a SQL-based query)}
}
\description{
Get a query function that will run a specific TPC-H query
}
