% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-file.R
\name{get_read_function}
\alias{get_read_function}
\title{Get a reader}
\usage{
get_read_function(file_type)
}
\arguments{
\item{file_type}{what file_type to read}
}
\value{
the read function to use
}
\description{
Get a reader
}
