% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{get_source_attr}
\alias{get_source_attr}
\title{Get source attributes}
\usage{
get_source_attr(file, attr)
}
\arguments{
\item{file}{the file to get attributes for}

\item{attr}{the attribute to get}
}
\description{
Get source attributes
}
\keyword{internal}
