% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{get_sql_query_func}
\alias{get_sql_query_func}
\title{Get a SQL query}
\usage{
get_sql_query_func(con, query_num)
}
\arguments{
\item{con}{the connection to DuckDB to use to get the query (note: this is
not the connection that will be used to run the query)}

\item{query_num}{the query number to fetch the result for}
}
\value{
a function that accepts an argument \code{con} which will run
\code{DBI::dbGetQuery()} against.
}
\description{
Produces a function that can be queried against any DBI backend (e.g. DuckDB)
}
\keyword{internal}
