% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-write-file.R, R/ensure-format.R
\name{get_write_function}
\alias{get_write_function}
\title{Get a writer}
\usage{
get_write_function(format, compression, num_groups)

get_write_function(format, compression, num_groups)
}
\arguments{
\item{format}{format to write}

\item{compression}{compression to use}

\item{num_groups}{how many rows groups to create (if the format supports it)}
}
\value{
the write function to use

the write function to use
}
\description{
Get a writer

Get a writer
}
