% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-dependencies.R
\name{install_benchconnect}
\alias{install_benchconnect}
\title{Install benchconnect}
\usage{
install_benchconnect()
}
\description{
Install \href{https://github.com/conbench/conbench/tree/main/benchconnect}{benchconnect},
a utility for sending benchmark results to a Conbench server
}
