% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-dependencies.R
\name{install_datalogistik}
\alias{install_datalogistik}
\title{Install datalogistik}
\usage{
install_datalogistik()
}
\description{
Install \href{https://github.com/conbench/datalogistik}{datalogistik}, a utility
for generating, downloading, and converting datasets for benchmarking.
}
\details{
Only for interactive use.
}
