% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-dependencies.R
\name{install_pipx}
\alias{install_pipx}
\title{Install pipx}
\usage{
install_pipx()
}
\description{
Install \href{https://pypa.github.io/pipx/}{pipx}, a version of pip that installs
Python packages in isolated environments where they will always be available
regardless of which version of Python is presently on \verb{$PATH}. Especially
useful for installing packages designed to be used via CLIs.
}
