% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\docType{data}
\name{known_sources}
\alias{known_sources}
\title{Known data files}
\format{
An object of class \code{list} of length 9.
}
\usage{
known_sources
}
\description{
Known data files
}
\keyword{datasets}
