% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-format.R
\docType{data}
\name{knowns}
\alias{knowns}
\alias{known_compressions}
\alias{known_formats}
\title{Known formats and compressions}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 4.
}
\usage{
known_compressions

known_formats
}
\description{
These formats and compression algorithms are known to {arrowbench}. Not all of
them will work with all formats (in fact, parquet is the only one that
supports all of them).
}
\keyword{datasets}
