% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Thing to test for \code{NULL}-ness}

\item{b}{Thing to use if \code{a} is \code{NULL}}
}
\value{
\code{a} unless it's \code{NULL}, then \code{b}
}
\description{
Default value for NULL
}
