% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-placebo.R
\docType{data}
\name{placebo}
\alias{placebo}
\title{Placebo benchmark for testing}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
placebo
}
\description{
Placebo benchmark for testing
}
\section{Parameters}{

\itemize{
\item \code{duration} the duration for the benchmark to take
}
}

\keyword{datasets}
