% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-csv.R
\docType{data}
\name{read_csv}
\alias{read_csv}
\title{Benchmark CSV reading}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
read_csv
}
\description{
Benchmark CSV reading
}
\section{Parameters}{

\itemize{
\item \code{source} A CSV file path to read in
\item \code{reader} One of \code{c("arrow", "data.table", "vroom", "readr")}
\item \code{compression} One of \code{c("uncompressed", "gzip")}
\item \code{output_format} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
