% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-file.R
\docType{data}
\name{read_file}
\alias{read_file}
\title{Benchmark file reading}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
read_file
}
\description{
Benchmark file reading
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id, or a CSV(?) file path to read in
\item \code{format} One of \code{c("parquet", "feather", "fst")}
\item \code{compression} One of the values: uncompressed, snappy, zstd, gzip, lz4, brotli, lzo, bz2
\item \code{output} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
