% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-json.R
\docType{data}
\name{read_json}
\alias{read_json}
\title{Benchmark JSON reading}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
read_json
}
\description{
Benchmark JSON reading
}
\section{Parameters}{

\itemize{
\item \code{source} A JSON file path to read in
\item \code{reader} One of \code{c("arrow", "jsonlite", "ndjson", "jsonify", "RcppSimdJson")}
\item \code{compression} One of \code{c("uncompressed", "gzip")}
\item \code{output} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
