% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{read_source}
\alias{read_source}
\title{Read a known source}
\usage{
read_source(file, ...)
}
\arguments{
\item{file}{file to read}

\item{...}{extra arguments to pass}
}
\value{
the source
}
\description{
Read a known source
}
