% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-remote-dataset.R
\docType{data}
\name{remote_dataset}
\alias{remote_dataset}
\title{Remote (S3) dataset reading}
\format{
An object of class \code{Benchmark} of length 9.
}
\usage{
remote_dataset
}
\description{
Remote (S3) dataset reading
}
\keyword{datasets}
