% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_bm}
\alias{run_bm}
\title{Execute a benchmark run}
\usage{
run_bm(bm, ..., n_iter = 1, profiling = FALSE)
}
\arguments{
\item{bm}{\code{\link[=Benchmark]{Benchmark()}} object}

\item{...}{parameters passed to \code{bm$setup()}.}

\item{n_iter}{integer number of iterations to replicate each benchmark}

\item{profiling}{Logical: collect prof info? If \code{TRUE}, the result data will
contain a \code{prof_file} field, which you can read in with
\code{profvis::profvis(prof_input = file)}. Default is \code{FALSE}}
}
\description{
This is the function that gets called in the script that \code{\link[=run_one]{run_one()}} prepares.
You may call this function interactively, but you won't get the isolation
in a fresh R process that \code{run_one()} provides.
}
