% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sync_and_drop_caches}
\alias{sync_and_drop_caches}
\title{Attempt to drop disk caches}
\usage{
sync_and_drop_caches()
}
\value{
Logical; were caches cleared?
}
\description{
Attempts to drop disk caches. Currently only works on Linux. If clearing
fails, will set an option so it will not reattempt on future calls.
}
\keyword{internal}
