% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-table-to-df.R
\docType{data}
\name{table_to_df}
\alias{table_to_df}
\title{Benchmark for reading an Arrow table to a data.frame}
\format{
An object of class \code{Benchmark} of length 11.
}
\usage{
table_to_df
}
\description{
This flexes conversion to R data structures from Arrow data structures.
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id to use (it will be read in to a data.frame first)
}
}

\keyword{datasets}
