% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{tables_refed}
\alias{tables_refed}
\title{For extracting table names from TPC-H queries}
\usage{
tables_refed(query_func)
}
\arguments{
\item{query_func}{a function containing a dplyr pipeline}
}
\value{
all references inside of \code{input_func(...)}, collapsed
}
\description{
This searches a function for all references of \code{input_func(...)} and returns
the contents of \code{...}
}
