% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\docType{data}
\name{tpc_h}
\alias{tpc_h}
\title{Benchmark TPC-H queries}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
tpc_h
}
\description{
Benchmark TPC-H queries
}
\section{Parameters}{

\itemize{
\item \code{engine} One of \verb{c("arrow", "duckdb", "dplyr)}
\item \code{query_num} integer, 1-19
\item \code{scale} Scale factor to use for data generation
}
}

\keyword{datasets}
