% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpch-queries.R
\docType{data}
\name{tpc_h_queries}
\alias{tpc_h_queries}
\title{all queries take an input_func which is a function that will return a dplyr tbl
referencing the table needed.}
\format{
An object of class \code{list} of length 22.
}
\usage{
tpc_h_queries
}
\description{
all queries take an input_func which is a function that will return a dplyr tbl
referencing the table needed.
}
\keyword{internal}
