% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-tpch-source.R
\docType{data}
\name{tpch_tables}
\alias{tpch_tables}
\title{Table names for TPC-H benchmarks}
\format{
An object of class \code{character} of length 8.
}
\usage{
tpch_tables
}
\description{
Table names for TPC-H benchmarks
}
\keyword{internal}
