% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{URSA_I9_9960X_R_BENCHMARK_NAMES}
\alias{URSA_I9_9960X_R_BENCHMARK_NAMES}
\title{A vector of benchmark attribute names run on \verb{ursa-i9-9960x}}
\format{
An object of class \code{character} of length 6.
}
\usage{
URSA_I9_9960X_R_BENCHMARK_NAMES
}
\description{
A vector of benchmark attribute names run on \verb{ursa-i9-9960x}
}
\keyword{datasets}
