% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-format.R
\name{validate_format}
\alias{validate_format}
\alias{stop_if_not_valid_format}
\title{Validate format and compression combinations}
\usage{
validate_format(format = known_formats, compression = known_compressions)

stop_if_not_valid_format(format, compression)
}
\arguments{
\item{format}{the format of the file}

\item{compression}{the compression codec}
}
\value{
\code{TRUE} invisibly
}
\description{
For a given format + compression, determine if the combination is valid.
\code{validate_format()} returns a vector of \code{TRUE}/\code{FALSE} if the formats are
valid. \code{stop_if_not_valid_format()} will stop if any of the format + compressions
are not valid.
}
\keyword{internal}
