% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-write-csv.R
\docType{data}
\name{write_csv}
\alias{write_csv}
\title{Benchmark CSV writing}
\format{
An object of class \code{Benchmark} of length 11.
}
\usage{
write_csv
}
\description{
Benchmark CSV writing
}
\section{Parameters}{

\itemize{
\item \code{source} A CSV file path to write to
\item \code{writer} One of \code{c("arrow", "data.table", "vroom", "readr",)}
\item \code{input} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
