% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-write-file.R
\docType{data}
\name{write_file}
\alias{write_file}
\title{Benchmark file writing}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
write_file
}
\description{
Benchmark file writing
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id, or a CSV(?) file path to read in
\item \code{format} One of \code{c("parquet", "feather", "fst")}
\item \code{compression} One of the values: uncompressed, snappy, zstd, gzip, lz4
\item \code{input} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
