% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-array-to-vector.R
\docType{data}
\name{array_to_vector}
\alias{array_to_vector}
\title{Benchmark for reading an Arrow array to a vector}
\format{
An object of class \code{Benchmark} of length 9.
}
\usage{
array_to_vector
}
\description{
This flexes a lower level conversion to R data structures from Arrow data structures.
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id to use (it will be read in to a data.frame first)
\item \code{chunked_arrays} logical, should the arrays converted be \code{ChunkedArrays} or \code{Arrays}?
\item \code{exclude_nulls} logical, should any columns with any \code{NULL}s or \code{NA}s in them be removed?
\item \code{alt_rep} logical, should the altrep option be set? (\code{TRUE} to enable it, \code{FALSE} to disable)
}
}

\keyword{datasets}
