% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-R-to-arrow.R
\docType{data}
\name{df_to_table}
\alias{df_to_table}
\title{Benchmark file writing}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
df_to_table
}
\description{
Benchmark file writing
}
\section{Parameters}{

\itemize{
\item \code{source} A known-file id, or a CSV(?) file path to read in
\item \code{format} One of \code{c("parquet", "feather", "fst")}
\item \code{compression} One of \code{c("uncompressed", "snappy", "zstd")}
\item \code{input} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
