% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{ensure_format}
\alias{ensure_format}
\title{Ensure that a source has a specific format}
\usage{
ensure_format(
  name,
  format = c("csv", "parquet", "feather", "csv.gz"),
  compression = c("uncompressed", "snappy", "zstd", "gzip", "gz", "lz4")
)
}
\arguments{
\item{name}{name of the known source}

\item{format}{format to be ensured}

\item{compression}{compression to be ensured}
}
\value{
the file that was ensured to exist
}
\description{
Ensure that a source has a specific format
}
