% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{get_input_func}
\alias{get_input_func}
\title{Get an input function for a table}
\usage{
get_input_func(
  engine,
  scale_factor,
  query_id,
  format,
  compression = "uncompressed",
  con = NULL,
  memory_map = FALSE,
  chunk_size = NULL
)
}
\arguments{
\item{engine}{which engine to use}

\item{scale_factor}{what scale factor to reference}

\item{query_id}{which query is being used}

\item{format}{which format}

\item{compression}{which compression to use (default: "uncompressed")}

\item{con}{a connection}

\item{memory_map}{should the file be memory mapped? (only relevant for the "native" format with Arrow)}

\item{chunk_size}{what chunk_size should be used with the source files? (default: NULL, the default for the file format)}
}
\description{
This returns a function which will return a table reference with the specified
parameters
}
