% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-write-file.R, R/ensure-format.R
\name{get_write_function}
\alias{get_write_function}
\title{Get a writer}
\usage{
get_write_function(format, compression, chunk_size = NULL)

get_write_function(format, compression, chunk_size = NULL)
}
\arguments{
\item{format}{format to write}

\item{compression}{compression to use}

\item{chunk_size}{the size of chunks to write (default: NULL, the default for
the format)}

\item{file_type}{file_type to write}
}
\value{
the write function to use

the write function to use
}
\description{
Get a writer

Get a writer
}
