% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure times and memory usage}
\usage{
measure(..., profiling = FALSE)
}
\arguments{
\item{...}{An expression to run}

\item{profiling}{Logical: collect prof info? If \code{TRUE}, the result data will
contain a \code{prof_file} field, which you can read in with
\code{profvis::profvis(prof_input = file)}. Default is false}
}
\value{
A tibble of timings and memory usage
}
\description{
Measure times and memory usage
}
