% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-read-csv.R, R/bm-read-file.R
\docType{data}
\name{read_file}
\alias{read_file}
\title{Benchmark CSV reading}
\format{
An object of class \code{Benchmark} of length 7.

An object of class \code{Benchmark} of length 7.
}
\usage{
read_file

read_file
}
\description{
Benchmark CSV reading

Benchmark file reading
}
\section{Parameters}{

\itemize{
\item \code{source} A CSV file path to read in
\item \code{reader} One of \code{c("arrow", "data.table", "vroom", "readr")}
\item \code{compression} One of \code{c("uncompressed", "gzip")}
\item \code{output} One of \code{c("arrow_table", "data_frame")}
}


\itemize{
\item \code{source} A known-file id, or a CSV(?) file path to read in
\item \code{format} One of \code{c("parquet", "feather", "fst")}
\item \code{compression} One of \code{c("uncompressed", "snappy", "zstd")}
\item \code{output} One of \code{c("arrow_table", "data_frame")}
}
}

\keyword{datasets}
