% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{run.BenchmarkDataFrame}
\title{Run an object}
\usage{
run(x, ...)

\method{run}{BenchmarkDataFrame}(x, ...)
}
\arguments{
\item{x}{An S3 classed object to run}

\item{...}{Additional arguments passed through to methods. For
\code{run.BenchmarkDataFrame}, passed through to \code{\link[=get_default_parameters]{get_default_parameters()}} (when
parameters are not specified) and \code{\link[=run_benchmark]{run_benchmark()}}.}
}
\value{
A modified object containing run results. For \code{run.BenchmarkDataFrame},
a \code{results} list column is appended.
}
\description{
Run an object
}
