% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_one}
\alias{run_one}
\title{Run a Benchmark with a single set of parameters}
\usage{
run_one(
  bm,
  ...,
  n_iter = 1,
  dry_run = FALSE,
  profiling = FALSE,
  progress_bar = NULL,
  read_only = FALSE,
  output_script = TRUE,
  test_packages = NULL
)
}
\arguments{
\item{bm}{\code{\link[=Benchmark]{Benchmark()}} object}

\item{...}{parameters passed to \code{bm$setup()}.}

\item{n_iter}{integer number of iterations to replicate each benchmark}

\item{dry_run}{logical: just return the R source code that would be run in
a subprocess? Default is \code{FALSE}, meaning that the benchmarks will be run.}

\item{profiling}{Logical: collect prof info? If \code{TRUE}, the result data will
contain a \code{prof_file} field, which you can read in with
\code{profvis::profvis(prof_input = file)}. Default is \code{FALSE}}

\item{progress_bar}{a \code{progress} object to update progress to (default \code{NULL})}

\item{read_only}{this will only attempt to read benchmark files and will not
run any that it cannot find.}

\item{output_script}{logical: output the rscript used to run the benchmark as a
json file}

\item{test_packages}{a character vector of packages that the benchmarks test (default \code{NULL})}
}
\value{
A \code{arrowbench_result}: a \code{list} containing "params" and either
"result" or "error".
}
\description{
Run a Benchmark with a single set of parameters
}
