% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\docType{data}
\name{tpc_h}
\alias{tpc_h}
\title{Benchmark TPC-H queries}
\format{
An object of class \code{Benchmark} of length 8.
}
\usage{
tpc_h
}
\description{
Benchmark TPC-H queries
}
\section{Parameters}{

\itemize{
\item \code{engine} One of \code{c("arrow", "duckdb")}
\item \code{query_num} integer, 1-22
\item \code{format} One of \code{c("parquet", "feather", "native")}
\item \code{scale} Scale factor to use for data generation (e.g. 0.1, 1, 10, 100)
\item \code{mem_map} Should memory mapping be used when reading a file in? (only
applicable to arrow, native. \code{FALSE} will result in the file being explicitly
read into memory before the benchmark)
}
}

\keyword{datasets}
