% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm-tpc-h.R
\name{tpch_answer}
\alias{tpch_answer}
\title{Get a TPC-H answer}
\usage{
tpch_answer(
  scale_factor,
  query_id,
  source = c("arrowbench", "duckdb"),
  datasource = c("datalogistik", "duckdb")
)
}
\arguments{
\item{scale_factor}{scale factor (possible values: \code{c(0.01, 0.1, 1, 10)})}

\item{query_id}{Id of the query (possible values: 1-22)}

\item{source}{source of the answer (default: "arrowbench"), "duckdb" can
return answers for scale_factor 1.}

\item{datasource}{whcih data source should the answers be from (defualt:
"datalogistik")? Some data sources have slightly different data in some
columns (possible other value: "duckdb")}
}
\value{
the answer, as a data.frame
}
\description{
Get a TPC-H answer
}
