% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-format.R
\name{validate_format}
\alias{validate_format}
\title{Validate format and compression combinations}
\usage{
validate_format(format = known_formats, compression = known_compressions)
}
\arguments{
\item{format}{the format of the file}

\item{compression}{the compression codec}
}
\value{
\code{TRUE} invisibly
}
\description{
For a given format + compression, this will \code{stop} if the combo is not valid
and will return \code{TRUE} invisibly if it is.
}
\keyword{internal}
