% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure-source.R
\name{validate_format}
\alias{validate_format}
\title{Validate format and compression combinations}
\usage{
validate_format(format, compression)
}
\arguments{
\item{format}{the format of the file}

\item{compression}{the compression codec}
}
\value{
\code{TRUE} invisibly
}
\description{
For a given format + compression, determine if the combination is valid.
\code{validate_format()} returns a vector of \code{TRUE}/\code{FALSE} if the formats are
valid.
}
\keyword{internal}
